+++ unreal_plugin/Source/UnrealMCPEditor/Public/MCPChatPanel.h
@@ -7,6 +7,7 @@
 class SScrollBox;
 class SMultiLineEditableTextBox;
 class STextBlock;
+class SVerticalBox;
 
 class SMCPChatPanel : public SCompoundWidget
 {
@@ -28,6 +29,13 @@ private:
 
 	FReply HandleSendClicked();
 	FReply HandleClearClicked();
+	FReply HandleComposerKeyDown(const FGeometry& MyGeometry, const FKeyEvent& InKeyEvent);
+	FReply HandleCopyClicked(FString Message) const;
+	FReply HandleRerunClicked(FString Message, FString Sender);
+	FReply HandleOpenLogClicked();
+	FReply HandleRevealAssetClicked(FString Message);
+	virtual FReply OnDragOver(const FGeometry& MyGeometry, const FDragDropEvent& DragDropEvent) override;
+	virtual FReply OnDrop(const FGeometry& MyGeometry, const FDragDropEvent& DragDropEvent) override;
 	bool HandlePollTick(float DeltaTime);
 
 	void LoadHistory();
@@ -37,9 +45,21 @@ private:
 
 	void AddMessage(const FChatMessage& ChatMessage);
 	void RebuildMessageList();
+	TSharedRef<SWidget> BuildMessageWidget(const FChatMessage& ChatMessage);
+	TSharedRef<SWidget> BuildMarkdownMessageBody(const FChatMessage& ChatMessage);
+	void AddMarkdownBlocks(const FString& MarkdownText, const FString& MessageId, TSharedRef<SVerticalBox> BodyBox);
+	void AppendStreamingDelta(const FString& MessageId, const FString& Sender, const FString& Delta, bool bDone);
+	bool ApplySseLine(const FString& Line);
+	void InsertComposerText(const FString& Text);
 	void SetStatus(const FText& Text, const FSlateColor& Color);
 	void UpdateLastAgentTimestamp(const TArray<FChatMessage>& Messages);
 
+	FString MakeLocalMessageId() const;
+	FString NormaliseSender(const FString& Sender) const;
+	FText GetSenderLabel(const FString& Sender) const;
+	FSlateColor GetMessageColor(const FString& Sender) const;
+	FString BuildDropReference(const TSharedPtr<class FDragDropOperation>& Operation) const;
+	FString ExtractFirstAssetReference(const FString& Message) const;
 	FString BuildServerUrl(const FString& PathAndQuery) const;
 	FString MakeCurrentTimestamp() const;
 	TSharedRef<class IHttpRequest, ESPMode::ThreadSafe> MakeJsonRequest(const FString& Url, const FString& Verb) const;
@@ -54,6 +74,7 @@ private:
 	TSharedPtr<SScrollBox> MessageScrollBox;
 	TSharedPtr<SMultiLineEditableTextBox> MessageInput;
 	TSharedPtr<STextBlock> StatusText;
+	TMap<FString, TSharedPtr<STextBlock>> StreamingMessageTextBlocks;
 
 	FTSTicker::FDelegateHandle PollTickerHandle;
 };